src/store/reducers/budget.reducer.ts
import { LOAD_BUDGET_COMPLETE, UPDATE_BUDGET, ADD_BUDGET_COMPLETE, REMOVE_BUDGET_COMPLETE } from './../actions/actions';
import { Budget } from './../models/interfaces';

export function BudgetReducer(state: Budget[] = [], action: any) {
  switch (action.type) {
    case LOAD_BUDGET_COMPLETE:
      return action.payload;
    case ADD_BUDGET_COMPLETE:
      return [...state, action.payload];
    case UPDATE_BUDGET:
      return state.map(budget => {
        if (action.payload.id === budget.id) {
          return { ...budget, ...action.payload};
        }

        return budget;
      });
    case REMOVE_BUDGET_COMPLETE:
      return state.filter(budget => budget.id !== action.payload);
    default:
      return state;
  }
}
